pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "expenses"
version = "1.0.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "A package for managing and visualizing financial expense data"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "dash>=2.14.2",
    "pandas>=2.1.4",
    "plotly>=5.18.0",
    "pyyaml>=6.0.1",
    "openai>=1.6.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/expenses"
Repository = "https://github.com/yourusername/expenses.git"
Issues = "https://github.com/yourusername/expenses/issues"

[project.scripts]
expenses-classify = "expenses.classifier:main"
expenses-app = "app:main"

[tool.setuptools]
packages = ["expenses"]

[tool.setuptools.package-data]
expenses = ["config/*.yaml"]
